Streamlit/requirements.txt
streamlit==1.33.0
openai==0.28.0
requests==2.31.0
pypdf==3.16.2
python-dotenv==1.0.0